Introduce `DelayedDrop` and remove IO operations from `Drop`
This supports the goal of removing `Drop` entirely so we can move to `async`
functions (https://github.com/core-json/core-json/issues/8). This removes the
primary hurdle, the lack of `AsyncDrop` stablization. Given how annoying this
was, I perfectly understand why an asynchronous `Drop` hasn't been stablized.

This is merged as a subset of the async branch as it seems to be non-trivially
faster (5-10%).